requirements.txt
google-api-python-client==2.98.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
pandas==2.1.0
python-dateutil==2.8.2
tqdm==4.66.1